# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис предоставляет функционал для работы с сегментами и пользователями для целей аналитики",
        "title": "Сервис по работе с сегментами",
        "contact": {
            "name": "Грищенко Владимир",
            "email": "vladimirxsky@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/reports": {
            "post": {
                "description": "Возвращает csv-строку, представляющую собой csv-отчёт,\nсодержащий столбцы `user_id`, `segment_name`, `start_date`,\n`end_date`, обозначающие идентификатор пользователя,\nнаименование сегмента, дату добавления пользователя в сегмент и\nдату выхода пользователя из сегмента соответственно. Строки отчёта\nотсортированы в порядке возрастания по дате добавления пользователя в сегмент.\n\nВ результате выполнения запроса формируется файл и устанавливается заголовок ответа\n`Content-Disposition`, поэтому результат выполнения запроса необходимо скачать.",
                "tags": [
                    "reports"
                ],
                "summary": "Получить отчёт в формате csv",
                "responses": {
                    "200": {
                        "description": "Структура, содержащая дату формирования отчёта и отчёт в виде csv-строки",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.MakeReportResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrReportValidationError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/segments": {
            "get": {
                "description": "Возвращает список всех сегментов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "Получить список всех сегментов",
                "responses": {
                    "200": {
                        "description": "Список всех сегментов",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.GetAllSegmentsResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrSegmentValidationError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт сегмент на основе информации в теле запроса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "Создать сегмент",
                "parameters": [
                    {
                        "description": "Структура с информацией о создаваемом сегменте",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_service.SegmentCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Наименование созданного сегмента",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrSegmentValidationError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/segments/{name}": {
            "get": {
                "description": "Возвращает информацию о сегменте с указанным именем",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "Получить сегмент с указанным именем",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Наименование сегмента",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сегмент с указанным именем",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.GetSegmentByNameResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrSegmentValidationError"
                        }
                    },
                    "404": {
                        "description": "Сегмент с указанным именем не был найден",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrSegmentNotFound"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет сегмент с указанным именем из системы.\nВ случае, если на момент совершения запроса в этот\nсегмент входят какие-либо пользователи, они автоматически\nвыйдут из данного сегмента.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "Удалить сегмент с указанным именем",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Наименование сегмента",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успехе",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.DeleteSegmentByNameResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrSegmentValidationError"
                        }
                    },
                    "404": {
                        "description": "Сегмент с указанным именем не был найден",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrSegmentNotFound"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "description": "Возвращает список абсолютно всех пользователей",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получить список всех пользователей",
                "responses": {
                    "200": {
                        "description": "Список всех пользователей",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.GetAllUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserValidationError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт пользователя на основе информации в теле запроса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создать пользователя",
                "parameters": [
                    {
                        "description": "Структура с информацией о создаваемом пользователе",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_service.UserCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Идентификатор созданного пользователя",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.UserCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserValidationError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "avito-rest-api_internal_entity.Segment": {
            "type": "object",
            "properties": {
                "is_deleted": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "AVITO_MUSIC_SERVICE"
                },
                "segment_id": {
                    "type": "integer",
                    "example": 43
                }
            }
        },
        "avito-rest-api_internal_entity.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 26
                },
                "is_deleted": {
                    "type": "boolean",
                    "example": false
                },
                "lastname": {
                    "type": "string",
                    "example": "Иванов"
                },
                "name": {
                    "type": "string",
                    "example": "Михаил"
                },
                "sex": {
                    "description": "Пол, 0 - мужской, 1 - женский",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 0
                },
                "sex_text": {
                    "type": "string",
                    "enum": [
                        "мужской",
                        "женский"
                    ],
                    "example": "мужской"
                },
                "user_id": {
                    "type": "integer",
                    "example": 16
                }
            }
        },
...
        "avito-rest-api_internal_error.ErrInternalServerError": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "originErrorText": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "avito-rest-api_internal_error.ErrReportValidationError": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "originErrorText": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "avito-rest-api_internal_error.ErrSegmentNotFound": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "originErrorText": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "avito-rest-api_internal_error.ErrSegmentValidationError": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "originErrorText": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "avito-rest-api_internal_error.ErrUserValidationError": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "originErrorText": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "avito-rest-api_internal_service.SegmentCreateInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "Имя сегмента",
                    "type": "string",
                    "example": "AVITO_MUSIC_SERVICE"
                },
                "percentage": {
                    "description": "Необязательное поле, процент пользователей, которое автоматически войдёт в сегмент при его создании",
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 57
                }
            }
        },
        "avito-rest-api_internal_service.UserCreateInput": {
            "type": "object",
            "required": [
                "age",
                "lastname",
                "name",
                "sex"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 26
                },
                "lastname": {
                    "type": "string",
                    "example": "Иванов"
                },
                "name": {
                    "type": "string",
                    "example": "Михаил"
                },
                "sex": {
                    "description": "Пол, 0 - мужской, 1 - женский",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 0
                }
            }
        },
...
        "internal_controller_http_v1.CreateResponse": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "AVITO_MUSIC_SERVICE"
                }
            }
        },
        "internal_controller_http_v1.DeleteSegmentByNameResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "successfully deleted segment \"AVITO_MUSIC_SERVICE\""
                }
            }
        },
...
        "internal_controller_http_v1.GetAllSegmentsResponse": {
            "type": "object",
            "properties": {
                "segments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/avito-rest-api_internal_entity.Segment"
                    }
                }
            }
        },
        "internal_controller_http_v1.GetAllUsersResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/avito-rest-api_internal_entity.User"
                    }
                }
            }
        },
...
        "internal_controller_http_v1.GetSegmentByNameResponse": {
            "type": "object",
            "properties": {
                "segment": {
                    "$ref": "#/definitions/avito-rest-api_internal_entity.Segment"
                }
            }
        },
...
        "internal_controller_http_v1.MakeReportResponse": {
            "type": "object",
            "properties": {
                "report": {
                    "description": "Отчёт в виде csv-строки с разделителями \",\" и символом перехода на новую строку \"\\n\"",
                    "type": "string"
                },
                "report_date": {
                    "description": "Дата формирования отчёта",
                    "type": "string"
                }
            }
        },
        "internal_controller_http_v1.UserCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 26
                }
            }
        }
    }
}